+++ game/ato/flightstate/inflight.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from collections import deque
 from datetime import datetime, timedelta
+import logging
 from typing import TYPE_CHECKING
 
 from dcs import Point
@@ -13,6 +14,7 @@ from game.ato.flightstate.flightstate import FlightState
 from game.ato.flightwaypoint import FlightWaypoint
 from game.ato.flightwaypointtype import FlightWaypointType
 from game.ato.starttype import StartType
+from game.settings.settings import FastForwardStopCondition
 from game.utils import Distance, LBS_TO_KG, Speed, pairwise
 
 if TYPE_CHECKING:
@@ -164,3 +166,17 @@ class InFlight(FlightState, ABC):
     @property
     def spawn_type(self) -> StartType:
         return StartType.IN_FLIGHT
+
+    def should_halt_sim(self) -> bool:
+        if (
+            self.flight.client_count > 0
+            and self.settings.fast_forward_stop_condition
+            == FastForwardStopCondition.PLAYER_AT_IP
+            and self.is_at_ip
+        ):
+            logging.info(
+                f"Interrupting simulation because {self.flight} has players and has "
+                "reached IP"
+            )
+            return True
+        return False